requirements.txt
motor
pymongo
aiohttp
python-dotenv